+++ tests/infrastructure/test_rana_dataset_gateway.py
@@ -4,7 +4,7 @@ from pydantic import AnyHttpUrl
 from pytest import fixture, raises
 
 from rana_process_sdk.domain import DoesNotExist, Json
-from rana_process_sdk.domain.dataset import RanaDataset, ResourceIdentifier
+from rana_process_sdk.domain.dataset import DatasetLink, RanaDataset, ResourceIdentifier
 from rana_process_sdk.infrastructure import RanaApiProvider, RanaDatasetGateway
 
 
@@ -26,6 +26,18 @@ def rana_dataset_response() -> Json:
             "default": "Titel",
         },
         "resourceIdentifier": [{"code": "LizardId", "link": "https://lizard/rasters"}],
+        "link": [
+            {
+                "protocol": "OGC:WCS",
+                "urlObject": {"default": "https://some/wcs?version=2.0.1"},
+                "nameObject": {"default": "dtm_05m"},
+            },
+            {
+                "protocol": "INSPIRE Atom",
+                "urlObject": {},
+                "nameObject": {},
+            },
+        ],
     }
 
 
@@ -41,6 +53,14 @@ def test_get(gateway: RanaDatasetGateway, provider: Mock, rana_dataset_response:
             ResourceIdentifier(
                 code="LizardId", link=AnyHttpUrl("https://lizard/rasters")
             )
+        ],
+        links=[
+            DatasetLink(
+                protocol="OGC:WCS",
+                name="dtm_05m",
+                url=AnyHttpUrl("https://some/wcs?version=2.0.1"),
+            ),
+            DatasetLink(protocol="INSPIRE Atom", name=None, url=None),
         ],
     )
     provider.job_request.assert_called_once_with("GET", "datasets/DatasetId")